simulation update
